fixing database problems
- fixed the parsing errors with the new excel sheet format for seeding the db
- removed age column on google sheet, and changed to dob_year instead of streamer_dob
- changed is_mature to mature_stream
changed image_link to logo
- renamed calculate_streamer_2.js to calculate_streamer.js, removing the old version
